Restructured Tests - Split Integration test into 'api' and 'security' tests
