+++ src/tests/unit-test/util.test.ts
@@ -1,4 +1,4 @@
-import { isEmpty } from "../util/data-service-util";
+import { isEmpty } from "../../util/data-service-util";
 
 test("return true if an object is empty", () => {
   const isEmptyObject: Boolean = isEmpty({});
